% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_conditions.R
\name{data_dependent_conditions}
\alias{data_dependent_conditions}
\title{Evaluate data-dependent conditions}
\usage{
data_dependent_conditions(
  input,
  conf_list,
  conditions_into_input,
  input_id = NULL,
  data_server = NULL
)
}
\arguments{
\item{input}{A list of inputs.}

\item{conf_list}{A list of configuration for each input.}

\item{conditions_into_input}{A list of conditions to evaluate for the input.}

\item{input_id}{A character string specifying the ID of the input to evaluate. Defaults to NULL.}

\item{data_server}{A list of data frames, each named with a table_id.}
}
\value{
This function returns the list of configuration for the input after evaluating the data
}
\description{
This function evaluates the conditions that depend on the data stored in the database for a given input.
}
