% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_conditions.R
\name{eval_conditions}
\alias{eval_conditions}
\title{Evaluate a Condition}
\usage{
eval_conditions(element, condition, comparison_value)
}
\arguments{
\item{element}{The first input value. The subject of the condition.}

\item{condition}{The condition to evaluate. Should be a string matching one of the available conditions.}

\item{comparison_value}{The second input value. The value against which the condition is evaluated.}
}
\value{
The result of the evaluation. The return type depends on the condition.
}
\description{
This function takes two values and a condition, and evaluates the condition on the values.
}
\examples{
eval_conditions(5, "equals", 5)  # Returns TRUE
eval_conditions(5, "greater_than", 3)  # Returns TRUE
eval_conditions(5, "less_than", 3)  # Returns FALSE

}
