% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_conditions.R
\name{evaluate_filter_into_input}
\alias{evaluate_filter_into_input}
\title{Evaluate filter conditions in an input}
\usage{
evaluate_filter_into_input(
  input,
  conf_list,
  conditions_into_input,
  input_id = NULL
)
}
\arguments{
\item{input}{A list of inputs.}

\item{conf_list}{A configuration list with filter conditions.}

\item{conditions_into_input}{A list of conditions to evaluate for the input.}

\item{input_id}{A character string specifying the ID of the input to evaluate. Defaults to NULL.}
}
\value{
This function returns the list of conditions for the input after evaluating the presence and value of its filter condition.
If a filter condition is present and its argument is non-null and non-empty, the corresponding filter condition is returned.
If the filter argument is null or empty, a default condition considering only unique values of the input is returned.
If a filter condition is not present, the function returns NULL.
}
\description{
This function evaluates whether an input should be filtered based on its filter conditions.
If a filter condition is present and its argument is non-null and non-empty, the input will be filtered accordingly.
If the filter argument is null or empty, the function sets a default condition to consider only unique values of the input.
If a filter condition is not present, the function does nothing.
}
