% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_condition}
\alias{get_condition}
\title{Get condition operator}
\usage{
get_condition(condition)
}
\arguments{
\item{condition}{A character string specifying the condition to translate into an operator. This can be one of the following: "equals", "in", "distinct", "is_na", "is_empty", "is_not_empty", "less_than", "greater_than".}
}
\value{
This function returns a string representing the operator corresponding to the condition provided.
}
\description{
This function takes a string specifying a condition and returns the corresponding operator.
}
\examples{
get_condition("equals") # returns "=="

}
