% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_recommender.R
\name{input_recommender}
\alias{input_recommender}
\title{Shiny Input Recommender}
\usage{
input_recommender(data, dic = NULL, inputs_from_json = NULL)
}
\arguments{
\item{data}{A data frame for which to recommend input controls.}

\item{dic}{An optional dictionary that provides information about the data frame. If not provided, the function
will generate one using \code{hdtable::create_dic()}.}
}
\value{
This function returns a list of recommended Shiny input controls. Each element of the list is a list
itself that contains the properties for the corresponding input control.
}
\description{
This function takes a data frame and optionally a dictionary, and recommends Shiny input controls based on
the properties of the columns in the data frame. The recommendation is based on a pre-defined set of rules.
}
\examples{
\dontrun{
# Create a data frame
df <- data.frame(
  a = 1:10,
  b = letters[1:10],
  c = as.Date("2023-01-01") + 1:10
)
# Use the function to get recommended input controls
recommended_inputs <- input_recommender(df)
}
}
