% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_conditions.R
\name{internal_conditions}
\alias{internal_conditions}
\title{Internal Condition Evaluation}
\usage{
internal_conditions(
  input,
  conf_list,
  sing_values = NULL,
  input_id = NULL,
  input_params,
  data_server = NULL
)
}
\arguments{
\item{input}{A list of inputs.}

\item{conf_list}{A list of configurations.}

\item{sing_values}{The reactive values.}

\item{input_id}{The id of the input.}

\item{input_params}{Parameters of the input.}

\item{data_server}{The data server.}
}
\value{
The configuration list.
}
\description{
This function evaluates whether the condition is a reactive value, a database access or
a list of conditions. The 'sing_values' parameter indicates the name of the reactive value
that contains the specification of the desired value.
}
