% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perform_operations}
\alias{perform_operations}
\title{Perform operations on a data frame}
\usage{
perform_operations(df, operations)
}
\arguments{
\item{df}{A data frame to perform the operations on.}

\item{operations}{A list of operations to perform. The operations should be named with the corresponding function names in dplyr package. For example, "filter", "select", "distinct", etc.}
}
\value{
This function returns a data frame that is the result of performing the operations on the original data frame.
}
\description{
This function performs a series of operations on a data frame based on a list of operation names.
}
\examples{
df <- data.frame(a = 1:5, b = 6:10)
operations <- list(filter = list(col = "a", condition = "greater_than", arg = 2), setNames = list(col = "a", col_name = "b"))
perform_operations(df, operations) # returns a data frame with a > 2

}
