% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{prepare_inputs}
\alias{prepare_inputs}
\title{Prepare Inputs}
\usage{
prepare_inputs(
  input,
  inputs_data,
  sing_values = NULL,
  input_params,
  input_names,
  data_server = NULL
)
}
\arguments{
\item{input}{A list of inputs.}

\item{inputs_data}{A list of data inputs.}

\item{sing_values}{The reactive values.}

\item{input_params}{Parameters of the input.}

\item{input_names}{The names of the inputs.}

\item{data_server}{The data server.}

\item{inputs_user}{A list of user inputs.}
}
\description{
This function is used to prepare the inputs by calling the internal_conditions function
and storing in a reactive values, the name of the input and its arguments.
}
