% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_conditions.R
\name{primary_conditions}
\alias{primary_conditions}
\title{Primary Conditions Evaluation}
\usage{
primary_conditions(input, input_id, condition, comparison_value)
}
\arguments{
\item{input}{A list of all the inputs in a Shiny application, usually provided within a reactive context.}

\item{input_id}{Character. The ID of the input to evaluate.}

\item{condition}{Character. The condition to evaluate. This can be any condition that can be evaluated by the
'eval_conditions' function, including 'is_null' which checks if the input value is NULL.}

\item{comparison_value}{The value to compare the input value to, if necessary. This is used by the 'eval_conditions'
function to evaluate conditions that require a comparison.}
}
\value{
This function returns TRUE if the input value meets the specified condition.
}
\description{
This function evaluates a set of conditions for a specific inputs. The purpose of this function is to determine
whether a specific condition is met for the input. The conditions are evaluated based on the 'condition' argument
and, if necessary, a 'comparison_value'.
}
\keyword{internal}
